refactor: modularize settings page and implement custom color picker for branding and roles
