+++ apps/frontend/package.json
@@ -41,6 +41,7 @@
     "next": "^16.0.7",
     "next-themes": "^0.4.6",
     "react": "^19.2.0",
+    "react-colorful": "^5.7.0",
     "react-day-picker": "^9.11.1",
     "react-dom": "^19.2.0",
     "react-hook-form": "^7.66.1",
